Generate Solana specific VAAs in Conductor.sol

+++ ethereum/contracts/icco/contributor/ContributorGetters.sol
@@ -47,11 +47,11 @@ contract ContributorGetters is ContributorState {
         return _state.provider.conductorContract;
     }
 
-    function sales(uint saleId_) public view returns (ContributorStructs.Sale memory sale){
+    function sales(uint256 saleId_) public view returns (ContributorStructs.Sale memory sale){
         return _state.sales[saleId_];
     }
 
-    function getSaleAcceptedTokenInfo(uint saleId_, uint tokenIndex) public view returns (uint16 tokenChainId, bytes32 tokenAddress, uint128 conversionRate){
+    function getSaleAcceptedTokenInfo(uint256 saleId_, uint256 tokenIndex) public view returns (uint16 tokenChainId, bytes32 tokenAddress, uint128 conversionRate){
         return (
             _state.sales[saleId_].acceptedTokensChains[tokenIndex],
             _state.sales[saleId_].acceptedTokensAddresses[tokenIndex],
@@ -59,45 +59,45 @@ contract ContributorGetters is ContributorState {
         );
     }
 
-    function getSaleTimeframe(uint saleId_) public view returns (uint256 start, uint256 end){
+    function getSaleTimeframe(uint256 saleId_) public view returns (uint256 start, uint256 end){
         return (
             _state.sales[saleId_].saleStart,
             _state.sales[saleId_].saleEnd
         );
     }
 
-    function getSaleStatus(uint saleId_) public view returns (bool isSealed, bool isAborted){
+    function getSaleStatus(uint256 saleId_) public view returns (bool isSealed, bool isAborted){
         return (
             _state.sales[saleId_].isSealed,
             _state.sales[saleId_].isAborted
         );
     }
 
-    function getSaleTokenAddress(uint saleId_) public view returns (bytes32 tokenAddress){
+    function getSaleTokenAddress(uint256 saleId_) public view returns (bytes32 tokenAddress){
         tokenAddress = _state.sales[saleId_].tokenAddress;
     }
 
-    function getSaleAllocation(uint saleId, uint tokenIndex) public view returns (uint256 allocation){
+    function getSaleAllocation(uint256 saleId, uint256 tokenIndex) public view returns (uint256 allocation){
         return _state.sales[saleId].allocations[tokenIndex];
     }
 
-    function getSaleExcessContribution(uint saleId, uint tokenIndex) public view returns (uint256 allocation){
+    function getSaleExcessContribution(uint256 saleId, uint256 tokenIndex) public view returns (uint256 allocation){
         return _state.sales[saleId].excessContributions[tokenIndex];
     }
 
-    function getSaleTotalContribution(uint saleId, uint tokenIndex) public view returns (uint256 contributed){
+    function getSaleTotalContribution(uint256 saleId, uint256 tokenIndex) public view returns (uint256 contributed){
         return _state.totalContributions[saleId][tokenIndex];
     }
 
-    function getSaleContribution(uint saleId, uint tokenIndex, address contributor) public view returns (uint256 contributed){
+    function getSaleContribution(uint256 saleId, uint256 tokenIndex, address contributor) public view returns (uint256 contributed){
         return _state.contributions[saleId][tokenIndex][contributor];
     }
 
-    function refundIsClaimed(uint saleId, uint tokenIndex, address contributor) public view returns (bool){
+    function refundIsClaimed(uint256 saleId, uint256 tokenIndex, address contributor) public view returns (bool){
         return _state.refundIsClaimed[saleId][tokenIndex][contributor];
     }
 
-    function allocationIsClaimed(uint saleId, uint tokenIndex, address contributor) public view returns (bool){
+    function allocationIsClaimed(uint256 saleId, uint256 tokenIndex, address contributor) public view returns (bool){
         return _state.allocationIsClaimed[saleId][tokenIndex][contributor];
     }
 }
